+++ src/main/scala/org/podval/tools/scalajs/ScalaJSTask.scala
@@ -1,10 +1,11 @@
 package org.podval.tools.scalajs
 
 import org.gradle.api.Task
+import org.opentorah.node.NodeExtension
 
 trait ScalaJSTask extends Task:
   setDescription(s"$flavour ScalaJS")
 
   protected def flavour: String
 
-  protected final def setUpNodeProject(): Unit = NodeExtension.get(getProject).setUpNodeProject(List("jsdom"))
+  protected final def setUpNodeProject(): Unit = NodeExtension.get(getProject).setUpProject(List("jsdom"))