+++ CHANGELOG.md
@@ -5,8 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.4.9] - 2023-05-31
 - tests: better test testing infrastructure
 - cleanup: ScalaJS tutorial and frameworks test projects are now written by the testing infrastructure
+- chore: Scala 3.3.0
+- chore: ScalaTest 3.2.16
+- chore: ScalaJS 1.13.1
+- chore: Gradle 8.1.1
+- chore: tracking OpenTorah Node changes
 
 ## [0.4.8] - 2023-03-20
 - feat: `node` extension